added the bar.html
