+++ frontend/js/services/barApi.js
@@ -1,4 +1,4 @@
-import { BASE_URL } from "./apiConfig";
+import { BASE_URL } from "./apiConfig.js";
 
 const BAR_UPDATES_URL = `${BASE_URL}/bar/updates`;
 const BAR_BUMP_URL = `${BASE_URL}/bar/tables`;
@@ -19,5 +19,5 @@ export async function bumpBarTicket(tableNumber) {
   const res = await fetch(`${BAR_BUMP_URL}/${tableNumber}/bump`, {
     method: "POST",
   });
-  if (!res.ok) throw new Error(`Failed to bump bar ticker: HTTP ${res.status}`);
+  if (!res.ok) throw new Error(`Failed to bump bar ticket: HTTP ${res.status}`);
 }
